+++ CHANGELOG.md
@@ -34,6 +34,8 @@ Notes
   - When demo mode is enabled, pressing the Add button with `demo-dl` or `demo-dl:0b` entered as the model name in the add model sheet will simulate a one-minute model download.
 - **Add an unload model button to the server screen inspector**
   - This allows you to quickly unload (free from memory) models currently loaded into Ollama server memory.
+- **Add a Keep Alive option to the chat and image generation screen inspectors**
+  - This allows you to specify how long the model is kept in memory on the Ollama server.
 
 ### Bug Fixes and Improvements
 - **Fix to display image model details correctly**